+++ inst/sql/users.sql
@@ -1,14 +1,16 @@
+-- start from 2024-07-20, columns related to organization structure will not be supported
 SELECT DISTINCT
-    vo.OrganizationUserId user_id,
-    vo.RealName user_name,
-    vo.Gender user_sex,
-    vo.Birthday user_dob,
-    vo.OrganizationName organization_name,
-    vo.GradeName grade_name,
-    vo.ClassName class_name
+    project_course_user.OrganizationUserId user_id
+    { columns }
 FROM
-    iquizoo_business_db.project_course_config pcc
-    INNER JOIN iquizoo_business_db.project_course_user pcu ON pcu.ProjectCourseConfigId = pcc.Id AND pcu.Deleted <>1 AND pcc.Deleted <> 1
-    INNER JOIN iquizoo_business_db.v_organizationuser vo ON vo.OrganizationUserId = pcu.OrganizationUserId
+    iquizoo_business_db.project_course_config
+    INNER JOIN iquizoo_business_db.project_course_user
+    ON project_course_user.ProjectCourseConfigId = project_course_config.Id
+        AND project_course_user.Deleted <> 1 AND project_course_config.Deleted <> 1
+    INNER JOIN iquizoo_user_db.organization_user
+    ON organization_user.Id = project_course_user.OrganizationUserId
+        AND organization_user.Deleted <> 1
+    INNER JOIN iquizoo_user_db.base_organization
+    ON base_organization.Id = organization_user.OrganizationId
 WHERE
-    pcc.Id = ?;
+    project_course_config.Id = ?;